[REFACTOR] search 리팩토링
