+++ src/main/java/com/example/sideproject/domain/search/entity/PostSearchType.java
@@ -1,0 +1,5 @@
+package com.example.sideproject.domain.search.entity;
+
+public enum PostSearchType {
+    project, pr
+}
