+++ src/main/java/com/example/sideproject/global/config/ElasticsearchConfig.java
@@ -2,7 +2,7 @@ package com.example.sideproject.global.config;
 
 import com.example.sideproject.domain.project.entity.Project;
 import com.example.sideproject.domain.project.repository.ProjectRepository;
-import com.example.sideproject.domain.search.dto.ProjectDocument;
+import com.example.sideproject.domain.search.entity.ProjectDocument;
 import com.example.sideproject.domain.search.repository.SearchProjectRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,9 +10,7 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Configuration
 @RequiredArgsConstructor
@@ -32,7 +30,7 @@ public class ElasticsearchConfig {
                     log.info("Elasticsearch에 인덱싱된 프로젝트가 없습니다. 초기 인덱싱을 시작합니다...");
                     List<Project> allProjects = projectRepository.findAll();
                     List<ProjectDocument> documents = allProjects.stream()
-                                    .map(ProjectDocument::fromProject)
+                            .map(ProjectDocument::new)
                             .toList();
                     searchProjectRepository.saveAll(documents);
                     log.info("초기 데이터 인덱싱 완료: {} 개의 프로젝트 인덱싱됨", allProjects.size());
